roachprod: move command init out of main
This PR refactors roachprod to move the commands out from the main.go to a separate package. Along with the commands, the flags are also moved. Separating the sub commands to a separate package will help us in better code maintenance and easier integration of these commands with other flows like drtprod. Also, initialisation outside of main can help us in making the commands unit testable.

Not that, with this PR, there is no change in functionality of roachprod.

The code will be further enhance for maintainability by separating the commands from commands.go into different categories (like cluster provisioning, datadog, prometheus) and make them separate go files.

Epic: none
Release note: None
