+++ pkg/BUILD.bazel
@@ -1182,8 +1182,8 @@ GO_TARGETS = [
     "//pkg/cmd/roachprod-microbench:roachprod-microbench_test",
     "//pkg/cmd/roachprod-stress:roachprod-stress",
     "//pkg/cmd/roachprod-stress:roachprod-stress_lib",
+    "//pkg/cmd/roachprod/cli:cli",
     "//pkg/cmd/roachprod/grafana:grafana",
-    "//pkg/cmd/roachprod/update:update",
     "//pkg/cmd/roachprod:roachprod",
     "//pkg/cmd/roachprod:roachprod_lib",
     "//pkg/cmd/roachtest/cluster:cluster",
