+++ pkg/cmd/roachprod/cli/BUILD.bazel
@@ -1,0 +1,45 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "cli",
+    srcs = [
+        "bash_completion.go",
+        "commands.go",
+        "flags.go",
+        "handlers.go",
+        "resgistry.go",
+        "update.go",
+        "util.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachprod/cli",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/cmd/roachprod/grafana",
+        "//pkg/roachprod",
+        "//pkg/roachprod/cloud",
+        "//pkg/roachprod/config",
+        "//pkg/roachprod/errors",
+        "//pkg/roachprod/fluentbit",
+        "//pkg/roachprod/install",
+        "//pkg/roachprod/opentelemetry",
+        "//pkg/roachprod/roachprodutil",
+        "//pkg/roachprod/ssh",
+        "//pkg/roachprod/ui",
+        "//pkg/roachprod/vm",
+        "//pkg/roachprod/vm/gce",
+        "//pkg/util/envutil",
+        "//pkg/util/flagutil",
+        "//pkg/util/timeutil",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_cockroachdb_errors//oserror",
+        "@com_github_fatih_color//:color",
+        "@com_github_spf13_cobra//:cobra",
+        "@com_google_cloud_go_storage//:storage",
+        "@org_golang_google_api//option",
+        "@org_golang_x_crypto//ssh",
+        "@org_golang_x_exp//maps",
+        "@org_golang_x_term//:term",
+        "@org_golang_x_text//language",
+        "@org_golang_x_text//message",
+    ],
+)
