+++ pkg/cmd/roachprod/cli/flags.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package main
+package cli
 
 import (
 	"fmt"
@@ -107,9 +107,9 @@ var (
 	fetchLogsTimeout time.Duration
 )
 
-func initFlags() {
+func initRootCmdFlags(rootCmd *cobra.Command) {
 	rootCmd.PersistentFlags().BoolVarP(&config.Quiet, "quiet", "q",
-		false || !term.IsTerminal(int(os.Stdout.Fd())), "disable fancy progress output")
+		!term.IsTerminal(int(os.Stdout.Fd())), "disable fancy progress output")
 	rootCmd.PersistentFlags().IntVarP(&config.MaxConcurrency, "max-concurrency", "", 32,
 		"maximum number of operations to execute on nodes concurrently, set to zero for infinite",
 	)
@@ -121,7 +121,9 @@ func initFlags() {
 		"use-shared-user", true,
 		fmt.Sprintf("use the shared user %q for ssh rather than your user %q",
 			config.SharedUser, config.OSUser.Username))
+}
 
+func initCreateCmdFlags(createCmd *cobra.Command) {
 	createCmd.Flags().DurationVarP(&createVMOpts.Lifetime,
 		"lifetime", "l", 12*time.Hour, "Lifetime of the cluster")
 	createCmd.Flags().BoolVar(&createVMOpts.SSDOpts.UseLocalSSD,
@@ -157,32 +159,43 @@ func initFlags() {
 	for _, providerName := range vm.AllProviderNames() {
 		if vm.Providers[providerName].Active() {
 			providerOptsContainer[providerName].ConfigureCreateFlags(createCmd.Flags())
+			// createCmd only accepts a single GCE project, as opposed to all the other
+			// commands.
+			providerOptsContainer[providerName].ConfigureClusterFlags(createCmd.Flags(), vm.SingleProject)
+		}
+	}
+}
 
+func initClusterFlagsForMultiProjects(
+	rootCmd *cobra.Command, excludeFromClusterFlagsMulti []*cobra.Command,
+) {
+	for _, providerName := range vm.AllProviderNames() {
+		if vm.Providers[providerName].Active() {
 			for _, cmd := range rootCmd.Commands() {
-				if cmd == createCmd {
-					// createCmd is handled below
+				excludeCmd := false
+				for _, c := range excludeFromClusterFlagsMulti {
+					if cmd == c {
+						excludeCmd = true
+						break
+					}
+				}
+				if excludeCmd {
 					continue
 				}
 				providerOptsContainer[providerName].ConfigureClusterFlags(cmd.Flags(), vm.AcceptMultipleProjects)
 			}
-
-			// createCmd only accepts a single GCE project, as opposed to all the other
-			// commands.
-			providerOptsContainer[providerName].ConfigureClusterFlags(createCmd.Flags(), vm.SingleProject)
-
-			// set up cluster cleanup flag for gcCmd
-			providerOptsContainer[providerName].ConfigureClusterCleanupFlags(gcCmd.Flags())
 		}
 	}
+}
 
+func initDestroyCmdFlags(destroyCmd *cobra.Command) {
 	destroyCmd.Flags().BoolVarP(&destroyAllMine,
 		"all-mine", "m", false, "Destroy all non-local clusters belonging to the current user")
 	destroyCmd.Flags().BoolVarP(&destroyAllLocal,
 		"all-local", "l", false, "Destroy all local clusters")
+}
 
-	extendCmd.Flags().DurationVarP(&extendLifetime,
-		"lifetime", "l", 12*time.Hour, "Lifetime of the cluster")
-
+func initListCmdFlags(listCmd *cobra.Command) {
 	listCmd.Flags().BoolVarP(&listCost,
 		"cost", "c", os.Getenv("ROACHPROD_COST_ESTIMATES") == "true",
 		"Show cost estimates",
@@ -195,29 +208,16 @@ func initFlags() {
 		"mine", "m", false, "Show only clusters belonging to the current user")
 	listCmd.Flags().StringVar(&listPattern,
 		"pattern", "", "Show only clusters matching the regex pattern. Empty string matches everything.")
+}
 
+func initAdminurlCmdFlags(adminurlCmd *cobra.Command) {
 	adminurlCmd.Flags().StringVar(&adminurlPath,
 		"path", "/", "Path to add to URL (e.g. to open a same page on each node)")
 	adminurlCmd.Flags().BoolVar(&adminurlIPs,
 		"ips", false, `Use Public IPs instead of DNS names in URL`)
+}
 
-	gcCmd.Flags().BoolVarP(&dryrun,
-		"dry-run", "n", dryrun, "dry run (don't perform any actions)")
-	gcCmd.Flags().StringVar(&config.SlackToken, "slack-token", "", "Slack bot token")
-
-	pgurlCmd.Flags().BoolVar(&external,
-		"external", false, "return pgurls for external connections")
-	for _, cmd := range []*cobra.Command{pgurlCmd, loadBalancerPGUrl} {
-		cmd.Flags().StringVar(&pgurlCertsDir,
-			"certs-dir", install.CockroachNodeCertsDir, "cert dir to use for secure connections")
-	}
-
-	for _, cmd := range []*cobra.Command{pgurlCmd, sqlCmd, loadBalancerPGUrl} {
-		cmd.Flags().StringVar(&authMode,
-			"auth-mode", install.DefaultAuthMode().String(), fmt.Sprintf("form of authentication to use, valid auth-modes: %v", maps.Keys(install.PGAuthModes)))
-		cmd.Flags().StringVar(&database, "database", "", "database to use")
-	}
-
+func initPprofCmdFlags(pprofCmd *cobra.Command) {
 	pprofCmd.Flags().DurationVar(&pprofOpts.Duration,
 		"duration", 30*time.Second, "Duration of profile to capture")
 	pprofCmd.Flags().BoolVar(&pprofOpts.Heap,
@@ -227,12 +227,9 @@ func initFlags() {
 	pprofCmd.Flags().IntVar(&pprofOpts.StartingPort,
 		"starting-port", 9000, "Initial port to use when opening pprof's HTTP interface")
 
-	ipCmd.Flags().BoolVar(&external,
-		"external", false, "return external IP addresses")
+}
 
-	runCmd.Flags().StringVarP(&extraSSHOptions,
-		"ssh-options", "O", "", "extra args to pass to ssh")
-
+func initStartCmdFlags(startCmd *cobra.Command) {
 	startCmd.Flags().IntVarP(&numRacks,
 		"racks", "r", 0, "the number of racks to partition the nodes into")
 	startCmd.Flags().StringArrayVarP(&startOpts.ExtraArgs,
@@ -251,53 +248,33 @@ func initFlags() {
 		"store-count", startOpts.StoreCount, "number of stores to start each node with")
 	startCmd.Flags().IntVar(&startOpts.AdminUIPort,
 		"admin-ui-port", startOpts.AdminUIPort, "port to serve the admin UI on")
+}
 
+func initStartInstanceCmdFlags(startInstanceCmd *cobra.Command) {
 	startInstanceCmd.Flags().StringVarP(&storageCluster, "storage-cluster", "S", "", "storage cluster")
 	_ = startInstanceCmd.MarkFlagRequired("storage-cluster")
 	startInstanceCmd.Flags().IntVar(&startOpts.SQLInstance,
 		"sql-instance", 0, "specific SQL/HTTP instance to connect to (this is a roachprod abstraction for separate-process deployments distinct from the internal instance ID)")
 	startInstanceCmd.Flags().StringVar(&startOpts.VirtualClusterLocation, "external-nodes", startOpts.VirtualClusterLocation, "if set, starts service in external mode, as a separate process in the given nodes")
+}
 
-	// Flags for processes that stop (kill) processes.
-	for _, stopProcessesCmd := range []*cobra.Command{stopCmd, stopInstanceCmd, deployCmd} {
-		// Cobra does not support reusing flags across multiple commands, especially
-		// if the defaults differ, so we need to supply different flags for the case
-		// where the defaults are different.
-		// See: https://github.com/spf13/cobra/issues/1398
-		sigPtr := &sig
-		waitPtr := &waitFlag
-		gracePeriodPtr := &gracePeriod
-		// deployCmd is a special case, because it is used to stop processes in a
-		// rolling restart, and we want to drain the nodes by default.
-		if stopProcessesCmd == deployCmd {
-			sigPtr = &deploySig
-			waitPtr = &deployWaitFlag
-			gracePeriodPtr = &deployGracePeriod
-		}
-		stopProcessesCmd.Flags().IntVar(sigPtr, "sig", *sigPtr, "signal to pass to kill")
-		stopProcessesCmd.Flags().BoolVar(waitPtr, "wait", *waitPtr, "wait for processes to exit")
-		stopProcessesCmd.Flags().IntVar(gracePeriodPtr, "grace-period", *gracePeriodPtr, "approx number of seconds to wait for processes to exit, before a forceful shutdown (SIGKILL) is performed")
-	}
-	deployCmd.Flags().DurationVar(&pause, "pause", pause, "duration to pause between node restarts")
-
+func initSyncCmdFlags(syncCmd *cobra.Command) {
 	syncCmd.Flags().BoolVar(&listOpts.IncludeVolumes, "include-volumes", false, "Include volumes when syncing")
 	syncCmd.Flags().StringArrayVarP(&listOpts.IncludeProviders, "clouds", "c",
 		make([]string, 0), "Specify the cloud providers when syncing. Important: Use this flag only if you are certain that you want to sync with a specific cloud. All DNS host entries for other clouds will be erased from the DNS zone.")
 
-	wipeCmd.Flags().BoolVar(&wipePreserveCerts, "preserve-certs", false, "do not wipe certificates")
+}
 
-	putCmd.Flags().BoolVar(&useTreeDist, "treedist", useTreeDist, "use treedist copy algorithm")
-
+func initStageCmdFlags(stageCmd *cobra.Command) {
 	stageCmd.Flags().StringVar(&stageOS, "os", "", "operating system override for staged binaries")
 	stageCmd.Flags().StringVar(&stageArch, "arch", "",
 		"architecture override for staged binaries [amd64, arm64, fips]; N.B. fips implies amd64 with openssl")
 
 	stageCmd.Flags().StringVar(&stageDir, "dir", "", "destination for staged binaries")
-	// N.B. stageURLCmd just prints the URL that stageCmd would use.
-	stageURLCmd.Flags().StringVar(&stageOS, "os", "", "operating system override for staged binaries")
-	stageURLCmd.Flags().StringVar(&stageArch, "arch", "",
-		"architecture override for staged binaries [amd64, arm64, fips]; N.B. fips implies amd64 with openssl")
 
+}
+
+func initLogsCmdFlags(logsCmd *cobra.Command) {
 	logsCmd.Flags().StringVar(&logsFilter,
 		"filter", "", "re to filter log messages")
 	logsCmd.Flags().Var(flagutil.Time(&logsFrom),
@@ -311,74 +288,45 @@ func initFlags() {
 	logsCmd.Flags().StringVar(&logsProgramFilter,
 		"logs-program", "^cockroach$", "regular expression of the name of program in log files to search")
 
+}
+
+func initStageURLCmdFlags(stageURLCmd *cobra.Command) {
+	// N.B. stageURLCmd just prints the URL that stageCmd would use.
+	stageURLCmd.Flags().StringVar(&stageOS, "os", "", "operating system override for staged binaries")
+	stageURLCmd.Flags().StringVar(&stageArch, "arch", "",
+		"architecture override for staged binaries [amd64, arm64, fips]; N.B. fips implies amd64 with openssl")
+}
+
+func initMonitorCmdFlags(monitorCmd *cobra.Command) {
 	monitorCmd.Flags().BoolVar(&monitorOpts.IgnoreEmptyNodes,
 		"ignore-empty-nodes", false,
 		"Automatically detect the (subset of the given) nodes which to monitor "+
 			"based on the presence of a nontrivial data directory.")
-
 	monitorCmd.Flags().BoolVar(&monitorOpts.OneShot,
 		"oneshot", false,
 		"Report the status of all targeted nodes once, then exit. The exit "+
 			"status is nonzero if (and only if) any node was found not running.")
+}
 
-	cachedHostsCmd.Flags().StringVar(&cachedHostsCluster,
-		"cluster", "", "print hosts matching cluster")
+func initUpdateCmdFlags(updateCmd *cobra.Command) {
+	updateCmd.Flags().BoolVar(&roachprodUpdateRevert, "revert", false, "restore roachprod to the previous version "+
+		"which would have been renamed to roachprod.bak during the update process")
+	updateCmd.Flags().StringVarP(&roachprodUpdateBranch, "branch", "b", "master", "git branch")
+	updateCmd.Flags().StringVarP(&roachprodUpdateOS, "os", "o", "linux", "OS")
+	updateCmd.Flags().StringVarP(&roachprodUpdateArch, "arch", "a", "amd64", "CPU architecture")
+}
 
-	grafanaStartCmd.Flags().StringVar(&grafanaConfig,
-		"grafana-config", "", "URI to grafana json config, supports local and http(s) schemes")
+func initGrafanaAnnotationCmdFlags(grafanaAnnotationCmd *cobra.Command) {
+	grafanaAnnotationCmd.Flags().StringArrayVar(&grafanaTags,
+		"tags", []string{}, "grafana annotation tags")
+	grafanaAnnotationCmd.Flags().StringVar(&grafanaDashboardUID,
+		"dashboard-uid", "", "grafana dashboard UID")
+	grafanaAnnotationCmd.Flags().Int64SliceVar(&grafanaTimeRange,
+		"time-range", []int64{}, "grafana annotation time range in epoch time")
 
-	grafanaStartCmd.Flags().StringVar(&grafanaArch, "arch", "",
-		"binary architecture override [amd64, arm64]")
+}
 
-	grafanaDumpCmd.Flags().StringVar(&grafanaDumpDir, "dump-dir", "",
-		"the absolute path to dump prometheus data to (use the contained 'prometheus-docker-run.sh' to visualize")
-
-	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogSite, "datadog-site", "us5.datadoghq.com",
-		"Datadog site to send telemetry data to (e.g., us5.datadoghq.com)")
-
-	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogAPIKey, "datadog-api-key", "",
-		"Datadog API key")
-
-	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogService, "datadog-service", "cockroachdb",
-		"Datadog service name for emitted logs")
-
-	fluentBitStartCmd.Flags().StringSliceVar(&fluentBitConfig.DatadogTags, "datadog-tags", []string{},
-		"Datadog tags as a comma-separated list in the format KEY1:VAL1,KEY2:VAL2")
-
-	opentelemetryStartCmd.Flags().StringVar(&opentelemetryConfig.DatadogSite, "datadog-site", "us5.datadoghq.com",
-		"Datadog site to send telemetry data to (e.g., us5.datadoghq.com)")
-
-	opentelemetryStartCmd.Flags().StringVar(&opentelemetryConfig.DatadogAPIKey, "datadog-api-key", "",
-		"Datadog API key")
-
-	opentelemetryStartCmd.Flags().StringSliceVar(&opentelemetryConfig.DatadogTags, "datadog-tags", []string{},
-		"Datadog tags as a comma-separated list in the format KEY1:VAL1,KEY2:VAL2")
-
-	sshKeysAddCmd.Flags().StringVar(&sshKeyUser, "user", config.OSUser.Username,
-		"the user to be associated with the new key",
-	)
-
-	jaegerStartCmd.Flags().StringVar(&jaegerConfigNodes, "configure-nodes", "",
-		"the nodes on which to set the relevant CRDB cluster settings")
-
-	sideEyeRootCmd.AddCommand(sideEyeInstallCmd)
-	sideEyeRootCmd.AddCommand(sideEyeSnapCmd)
-
-	initCmd.Flags().IntVar(&startOpts.InitTarget,
-		"init-target", startOpts.InitTarget, "node on which to run initialization")
-
-	snapshotDeleteCmd.Flags().BoolVar(&dryrun,
-		"dry-run", false, "dry run (don't perform any actions)")
-	snapshotCmd.AddCommand(snapshotCreateCmd)
-	snapshotCmd.AddCommand(snapshotListCmd)
-	snapshotCmd.AddCommand(snapshotDeleteCmd)
-	snapshotCmd.AddCommand(snapshotApplyCmd)
-
-	rootStorageCmd.AddCommand(rootStorageCollectionCmd)
-	rootStorageCollectionCmd.AddCommand(collectionStartCmd)
-	rootStorageCollectionCmd.AddCommand(collectionStopCmd)
-	rootStorageCollectionCmd.AddCommand(storageSnapshotCmd)
-	rootStorageCollectionCmd.AddCommand(collectionListVolumes)
+func initCollectionStartCmdFlags(collectionStartCmd *cobra.Command) {
 	collectionStartCmd.Flags().IntVarP(&volumeCreateOpts.Size,
 		"volume-size", "s", 10,
 		"the size of the volume in Gigabytes (GB) to create for each store. Note: This volume will be deleted "+
@@ -402,92 +350,143 @@ func initFlags() {
 		"the volume type that should be created. Provide a volume type that is connected to"+
 			" the provider chosen for the cluster. If no volume type is provided the provider default will be used. "+
 			"Note: This volume will be deleted once the VM is deleted.")
+}
 
-	updateCmd.Flags().BoolVar(&roachprodUpdateRevert, "revert", false, "restore roachprod to the previous version "+
-		"which would have been renamed to roachprod.bak during the update process")
-	updateCmd.Flags().StringVarP(&roachprodUpdateBranch, "branch", "b", "master", "git branch")
-	updateCmd.Flags().StringVarP(&roachprodUpdateOS, "os", "o", "linux", "OS")
-	updateCmd.Flags().StringVarP(&roachprodUpdateArch, "arch", "a", "amd64", "CPU architecture")
+func initGrafanaStartCmdFlags(grafanaStartCmd *cobra.Command) {
+	grafanaStartCmd.Flags().StringVar(&grafanaConfig,
+		"grafana-config", "", "URI to grafana json config, supports local and http(s) schemes")
 
-	for _, cmd := range []*cobra.Command{adminurlCmd, grafanaURLCmd, jaegerURLCmd} {
-		cmd.Flags().BoolVar(&urlOpen, "open", false, "Open the url in a browser")
-	}
+	grafanaStartCmd.Flags().StringVar(&grafanaArch, "arch", "",
+		"binary architecture override [amd64, arm64]")
+}
 
-	for _, cmd := range []*cobra.Command{createCmd, listCmd, destroyCmd} {
-		cmd.Flags().StringVarP(&username, "username", "u", os.Getenv("ROACHPROD_USER"),
-			"Username to run under, detect if blank")
-	}
+func initFluentBitStartCmdFlags(fluentBitStartCmd *cobra.Command) {
+	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogSite, "datadog-site", "us5.datadoghq.com",
+		"Datadog site to send telemetry data to (e.g., us5.datadoghq.com)")
 
-	for _, cmd := range []*cobra.Command{statusCmd, monitorCmd, startCmd,
-		stopCmd, runCmd, wipeCmd, reformatCmd, installCmd, putCmd, getCmd,
-		sqlCmd, pgurlCmd, adminurlCmd, ipCmd,
-	} {
-		cmd.Flags().BoolVar(
-			&ssh.InsecureIgnoreHostKey, "insecure-ignore-host-key", true, "don't check ssh host keys")
-	}
+	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogAPIKey, "datadog-api-key", "",
+		"Datadog API key")
 
-	for _, cmd := range []*cobra.Command{startCmd, startInstanceCmd} {
-		cmd.Flags().BoolVar(&startOpts.ScheduleBackups,
-			"schedule-backups", startOpts.ScheduleBackups,
-			"create a cluster backup schedule once the cluster has started (by default, "+
-				"full backup hourly and incremental every 15 minutes)")
-		cmd.Flags().StringVar(&startOpts.ScheduleBackupArgs,
-			"schedule-backup-args", startOpts.ScheduleBackupArgs,
-			"Recurrence and scheduled backup options specification")
-		cmd.Flags().Int64Var(&startOpts.NumFilesLimit, "num-files-limit", startOpts.NumFilesLimit,
-			"limit the number of files that can be created by the cockroach process")
-		cmd.Flags().IntVar(&startOpts.SQLPort,
-			"sql-port", startOpts.SQLPort, "port on which to listen for SQL clients")
-		cmd.Flags().BoolVar(&startOpts.EnableFluentSink,
-			"enable-fluent-sink", startOpts.EnableFluentSink,
-			"whether to enable the fluent-servers attribute in the CockroachDB logging configuration")
-	}
+	fluentBitStartCmd.Flags().StringVar(&fluentBitConfig.DatadogService, "datadog-service", "cockroachdb",
+		"Datadog service name for emitted logs")
 
-	for _, cmd := range []*cobra.Command{
-		startCmd, startInstanceCmd, statusCmd, stopCmd, runCmd,
-	} {
-		cmd.Flags().StringVar(&tag, "tag", "", "the process tag")
-	}
+	fluentBitStartCmd.Flags().StringSliceVar(&fluentBitConfig.DatadogTags, "datadog-tags", []string{},
+		"Datadog tags as a comma-separated list in the format KEY1:VAL1,KEY2:VAL2")
 
-	for _, cmd := range []*cobra.Command{
-		startCmd, putCmd, getCmd,
-	} {
-		cmd.Flags().BoolVar(new(bool), "scp", false, "DEPRECATED")
-		_ = cmd.Flags().MarkDeprecated("scp", "always true")
-	}
+}
 
-	for _, cmd := range []*cobra.Command{startCmd, startInstanceCmd, sqlCmd} {
-		cmd.Flags().StringVarP(&config.Binary,
-			"binary", "b", config.Binary, "the remote cockroach binary to use")
-	}
-	for _, cmd := range []*cobra.Command{startCmd, startInstanceCmd, stopInstanceCmd, createLoadBalancerCmd, sqlCmd, pgurlCmd, loadBalancerPGUrl, adminurlCmd, runCmd, jaegerStartCmd, grafanaAnnotationCmd, updateTargetsCmd, growCmd} {
-		// TODO(renato): remove --secure once the default of secure
-		// clusters has existed in roachprod long enough.
-		cmd.Flags().BoolVar(&secure,
-			"secure", secure, "use a secure cluster (DEPRECATED: clusters are secure by default; use --insecure to create insecure clusters.)")
-		cmd.Flags().BoolVar(&insecure,
-			"insecure", insecure, "use an insecure cluster")
-	}
-	for _, cmd := range []*cobra.Command{pgurlCmd, sqlCmd, adminurlCmd, stopInstanceCmd, createLoadBalancerCmd, loadBalancerPGUrl, loadBalancerIP, jaegerStartCmd} {
-		cmd.Flags().StringVar(&virtualClusterName,
-			"cluster", "", "specific virtual cluster to connect to")
-		cmd.Flags().IntVar(&sqlInstance,
-			"sql-instance", 0, "specific SQL/HTTP instance to connect to (this is a roachprod abstraction distinct from the internal instance ID)")
-	}
+func initOpentelemetryStartCmdFlags(opentelemetryStartCmd *cobra.Command) {
+	opentelemetryStartCmd.Flags().StringVar(&opentelemetryConfig.DatadogSite, "datadog-site", "us5.datadoghq.com",
+		"Datadog site to send telemetry data to (e.g., us5.datadoghq.com)")
 
-	for _, cmd := range []*cobra.Command{startCmd, listCmd, syncCmd} {
-		cmd.Flags().StringSliceVar(&config.DNSRequiredProviders,
-			"dns-required-providers", config.DefaultDNSRequiredProviders,
-			"the cloud providers that must be active to refresh DNS entries",
-		)
-	}
+	opentelemetryStartCmd.Flags().StringVar(&opentelemetryConfig.DatadogAPIKey, "datadog-api-key", "",
+		"Datadog API key")
 
-	grafanaAnnotationCmd.Flags().StringArrayVar(&grafanaTags,
-		"tags", []string{}, "grafana annotation tags")
-	grafanaAnnotationCmd.Flags().StringVar(&grafanaDashboardUID,
-		"dashboard-uid", "", "grafana dashboard UID")
-	grafanaAnnotationCmd.Flags().Int64SliceVar(&grafanaTimeRange,
-		"time-range", []int64{}, "grafana annotation time range in epoch time")
-	fetchLogsCmd.Flags().DurationVarP(&fetchLogsTimeout,
-		"timeout", "t", 5*time.Minute, "Timeout for fetching the logs from the cluster nodes")
+	opentelemetryStartCmd.Flags().StringSliceVar(&opentelemetryConfig.DatadogTags, "datadog-tags", []string{},
+		"Datadog tags as a comma-separated list in the format KEY1:VAL1,KEY2:VAL2")
+}
+
+func initGCCmdFlags(gcCmd *cobra.Command) {
+	gcCmd.Flags().BoolVarP(&dryrun,
+		"dry-run", "n", dryrun, "dry run (don't perform any actions)")
+	gcCmd.Flags().StringVar(&config.SlackToken, "slack-token", "", "Slack bot token")
+	// Allow each Provider to inject additional configuration flags
+	for _, providerName := range vm.AllProviderNames() {
+		if vm.Providers[providerName].Active() {
+			// set up cluster cleanup flag for gcCmd
+			providerOptsContainer[providerName].ConfigureClusterCleanupFlags(gcCmd.Flags())
+		}
+	}
+}
+
+func initFlagPgurlCertsDirForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&pgurlCertsDir,
+		"certs-dir", install.CockroachNodeCertsDir, "cert dir to use for secure connections")
+}
+
+func initFlagAuthModeNDatabaseForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&authMode,
+		"auth-mode", install.DefaultAuthMode().String(), fmt.Sprintf("form of authentication to use, valid auth-modes: %v", maps.Keys(install.PGAuthModes)))
+	cmd.Flags().StringVar(&database, "database", "", "database to use")
+}
+
+func initFlagOpenForCmd(cmd *cobra.Command) {
+	cmd.Flags().BoolVar(&urlOpen, "open", false, "Open the url in a browser")
+}
+
+func initFlagUsernameForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVarP(&username, "username", "u", os.Getenv("ROACHPROD_USER"),
+		"Username to run under, detect if blank")
+}
+
+// initFlagsStopProcessForCmd initializes Flags for processes that stop (kill) processes.
+func initFlagsStopProcessForCmd(
+	stopProcessesCmd *cobra.Command, sigPtr *int, waitPtr *bool, gracePeriodPtr *int,
+) {
+	// Cobra does not support reusing flags across multiple commands, especially
+	// if the defaults differ, so we need to supply different flags for the case
+	// where the defaults are different.
+	// See: https://github.com/spf13/cobra/issues/1398
+	stopProcessesCmd.Flags().IntVar(sigPtr, "sig", *sigPtr, "signal to pass to kill")
+	stopProcessesCmd.Flags().BoolVar(waitPtr, "wait", *waitPtr, "wait for processes to exit")
+	stopProcessesCmd.Flags().IntVar(gracePeriodPtr, "grace-period", *gracePeriodPtr, "approx number of seconds to wait for processes to exit, before a forceful shutdown (SIGKILL) is performed")
+}
+
+func initFlagsStartOpsForCmd(cmd *cobra.Command) {
+	cmd.Flags().BoolVar(&startOpts.ScheduleBackups,
+		"schedule-backups", startOpts.ScheduleBackups,
+		"create a cluster backup schedule once the cluster has started (by default, "+
+			"full backup hourly and incremental every 15 minutes)")
+	cmd.Flags().StringVar(&startOpts.ScheduleBackupArgs,
+		"schedule-backup-args", startOpts.ScheduleBackupArgs,
+		"Recurrence and scheduled backup options specification")
+	cmd.Flags().Int64Var(&startOpts.NumFilesLimit, "num-files-limit", startOpts.NumFilesLimit,
+		"limit the number of files that can be created by the cockroach process")
+	cmd.Flags().IntVar(&startOpts.SQLPort,
+		"sql-port", startOpts.SQLPort, "port on which to listen for SQL clients")
+	cmd.Flags().BoolVar(&startOpts.EnableFluentSink,
+		"enable-fluent-sink", startOpts.EnableFluentSink,
+		"whether to enable the fluent-servers attribute in the CockroachDB logging configuration")
+}
+
+func initFlagInsecureIgnoreHostKeyForCmd(cmd *cobra.Command) {
+	cmd.Flags().BoolVar(
+		&ssh.InsecureIgnoreHostKey, "insecure-ignore-host-key", true, "don't check ssh host keys")
+}
+
+func initFlagTagForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&tag, "tag", "", "the process tag")
+}
+
+func initFlagSCPForCmd(cmd *cobra.Command) {
+	cmd.Flags().BoolVar(new(bool), "scp", false, "DEPRECATED")
+	_ = cmd.Flags().MarkDeprecated("scp", "always true")
+}
+
+func initFlagBinaryForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVarP(&config.Binary,
+		"binary", "b", config.Binary, "the remote cockroach binary to use")
+}
+
+func initFlagInsecureForCmd(cmd *cobra.Command) {
+	// TODO(renato): remove --secure once the default of secure
+	// clusters has existed in roachprod long enough.
+	cmd.Flags().BoolVar(&secure,
+		"secure", secure, "use a secure cluster (DEPRECATED: clusters are secure by default; use --insecure to create insecure clusters.)")
+	cmd.Flags().BoolVar(&insecure,
+		"insecure", insecure, "use an insecure cluster")
+}
+
+func initFlagsClusterNSQLForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&virtualClusterName,
+		"cluster", "", "specific virtual cluster to connect to")
+	cmd.Flags().IntVar(&sqlInstance,
+		"sql-instance", 0, "specific SQL/HTTP instance to connect to (this is a roachprod abstraction distinct from the internal instance ID)")
+}
+
+func initFlagDNSRequiredProvidersForCmd(cmd *cobra.Command) {
+	cmd.Flags().StringSliceVar(&config.DNSRequiredProviders,
+		"dns-required-providers", config.DefaultDNSRequiredProviders,
+		"the cloud providers that must be active to refresh DNS entries",
+	)
 }
