--- pkg/cmd/roachprod/update/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "update",
-    srcs = [
-        "update.go",
-        "util.go",
-    ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachprod/update",
-    visibility = ["//visibility:public"],
-    deps = [
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_cockroachdb_errors//oserror",
-        "@com_google_cloud_go_storage//:storage",
-        "@org_golang_google_api//option",
-    ],
-)
